backend/requirements.txt
# Core Framework Dependencies
Flask>=3.1.3                    # Web framework
Flask-Cors>=6.0.2               # Handle Cross-Origin Resource Sharing
Flask-Limiter>=4.1.1            # Rate limiting for API endpoints
gunicorn>=25.2.0                # Production-grade WSGI server

# Database
SQLAlchemy>=2.0.48              # SQL toolkit and ORM
pydantic>=2.12.5                # Data validation using Python type annotations

# AI and Language Processing
langchain>=1.2.13               # Framework for LLM applications
langchain-classic>=1.0.3        # Backward-compat memory module
langchain-openai>=1.1.12        # OpenAI integration for LangChain
langchain-community>=0.4.1      # Community-supported modules for LangChain
langchain-core>=1.2.22          # Core abstractions
langchain-text-splitters>=1.1.1  # Text splitting utilities
openai>=2.30.0                  # OpenAI API client

# File Processing and Upload Handling
python-magic>=0.4.27            # File type detection
unstructured>=0.21.5            # Document parsing and text extraction
pypdf>=6.9.2                    # PDF file processing
python-docx>=1.2.0              # Microsoft Word document processing
Pillow>=12.1.1                  # Image processing library

# Environment and Security
python-dotenv>=1.2.2            # Environment variable management

# OCR Dependencies
pytesseract>=0.3.13             # Python wrapper for Google's Tesseract-OCR Engine

# System Requirements for OCR
# Before using OCR features, install tesseract-ocr on your system:
# Ubuntu/Debian: sudo apt-get install tesseract-ocr
# macOS: brew install tesseract
# Windows: Download and install from https://github.com/UB-Mannheim/tesseract/wiki

# Additional macOS Requirements (via Homebrew)
# brew install tesseract-lang  # for additional language support
# brew install leptonica      # for better image processing

# Notes:
# - Python 3.8+ is required
# - Some packages may require system-level dependencies (e.g., build tools)
# - For production deployment, consider using exact versions (==) instead of minimum versions (>=)
